float wrnfix

+++ wjson/serializer/float.hpp
@@ -37,7 +37,7 @@ public:
 
     ss << v ;
 
-    for (int i = 0; i < bufsize && buf[i]!='\0'; ++i)
+    for (size_t i = 0; i < bufsize && buf[i]!='\0'; ++i)
     {
       *(end++) = buf[i];
     }
